+++ mir/orbis/__main__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017 Allen Li
+# Copyright (C) 2018 Allen Li
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,3 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+import fire
+
+from mir.orbis import commands
+
+fire.Fire(commands)